@@ -1,74 +1,93 @@
 {
-  "name": "state-of-statemanagers",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points",
-    "nx": "nx",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "nx workspace-lint && ng lint",
-    "e2e": "ng e2e",
-    "affected:apps": "nx affected:apps",
-    "affected:libs": "nx affected:libs",
-    "affected:build": "nx affected:build",
-    "affected:e2e": "nx affected:e2e",
-    "affected:test": "nx affected:test",
-    "affected:lint": "nx affected:lint",
-    "affected:dep-graph": "nx affected:dep-graph",
-    "affected": "nx affected",
-    "format": "nx format:write",
-    "format:write": "nx format:write",
-    "format:check": "nx format:check",
-    "update": "ng update @nrwl/workspace",
-    "workspace-schematic": "nx workspace-schematic",
-    "dep-graph": "nx dep-graph",
-    "help": "nx help"
-  },
-  "private": true,
-  "dependencies": {
-    "@nrwl/angular": "9.5.1",
-    "@angular/animations": "^9.1.0",
-    "@angular/common": "^9.1.0",
-    "@angular/compiler": "^9.1.0",
-    "@angular/core": "^9.1.0",
-    "@angular/forms": "^9.1.0",
-    "@angular/platform-browser": "^9.1.0",
-    "@angular/platform-browser-dynamic": "^9.1.0",
-    "@angular/router": "^9.1.0",
-    "rxjs": "~6.5.4",
-    "zone.js": "^0.10.2",
-    "@nestjs/common": "^7.0.0",
-    "@nestjs/core": "^7.0.0",
-    "@nestjs/platform-express": "^7.0.0",
-    "reflect-metadata": "^0.1.13"
-  },
-  "devDependencies": {
-    "@angular/cli": "9.1.0",
-    "@nrwl/workspace": "9.5.1",
-    "@types/node": "~8.9.4",
-    "dotenv": "6.2.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~6.0.0",
-    "eslint": "6.8.0",
-    "typescript": "~3.8.3",
-    "prettier": "2.0.4",
-    "@nrwl/nest": "9.5.1",
-    "@angular/compiler-cli": "^9.1.0",
-    "@angular/language-service": "^9.1.0",
-    "@angular-devkit/build-angular": "0.901.0",
-    "codelyzer": "~5.0.1",
-    "jest-preset-angular": "8.1.2",
-    "@nrwl/jest": "9.5.1",
-    "jest": "25.2.3",
-    "@types/jest": "25.1.4",
-    "ts-jest": "25.2.1",
-    "cypress": "^4.1.0",
-    "@nrwl/cypress": "9.5.1",
-    "@nrwl/node": "9.5.1",
-    "@nestjs/schematics": "^7.0.0",
-    "@nestjs/testing": "^7.0.0"
-  }
+	"name": "state-of-statemanagers",
+	"version": "0.0.0",
+	"license": "MIT",
+	"scripts": {
+		"ng": "ng",
+		"nx": "nx",
+		"start": "ng serve",
+		"start:api": "nx serve api",
+		"start:ngrx": "nx serve ngrx-demo",
+		"start:akita": "nx serve akita-demo",
+		"build": "nx build",
+		"build:prod": "nx affected:build --all --prod",
+		"clean": "rimraf ./dist ./coverage ./.codacy-coverage ./typedocs",
+		"test": "nx run-many --all --target test",
+		"lint": "nx workspace-lint && ng lint",
+		"lint:ts": "tslint --project .",
+		"lint:es": "eslint .",
+		"lint:css": "stylelint --allow-empty-input --syntax=css ./**/*.css",
+		"lint:scss": "stylelint --allow-empty-input --syntax=scss ./**/*.scss",
+		"lint:md": "markdownlint --ignore 'CHANGELOG.md' --ignore-path '.gitignore' .",
+		"lint:rmd": "remark .",
+		"lint:format": "prettier --check .",
+		"e2e": "nx e2e",
+		"affected:apps": "nx affected:apps",
+		"affected:libs": "nx affected:libs",
+		"affected:build": "nx affected:build",
+		"affected:build:prod": "nx affected:build --prod",
+		"affected:e2e": "nx affected:e2e",
+		"affected:test": "nx affected:test",
+		"affected:lint": "nx affected:lint",
+		"affected:dep-graph": "nx affected:dep-graph",
+		"affected": "nx affected",
+		"format": "nx format:write",
+		"format:write": "nx format:write",
+		"format:check": "nx format:check",
+		"ncu": "ncu",
+		"ncu:u": "ncu -u",
+		"docs": "typedoc libs",
+		"update": "ng update @nrwl/workspace",
+		"workspace-schematic": "nx workspace-schematic",
+		"dep-graph": "nx dep-graph",
+		"help": "nx help",
+		"postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points",
+		"release": "standard-version",
+		"prepublishOnly": "yarn build"
+	},
+	"private": true,
+	"dependencies": {
+		"@nrwl/angular": "9.5.1",
+		"@angular/animations": "^10.0.3",
+		"@angular/common": "^10.0.3",
+		"@angular/compiler": "^10.0.3",
+		"@angular/core": "^10.0.3",
+		"@angular/forms": "^10.0.3",
+		"@angular/platform-browser": "^10.0.3",
+		"@angular/platform-browser-dynamic": "^10.0.3",
+		"@angular/router": "^10.0.3",
+		"rxjs": "~6.6.0",
+		"zone.js": "^0.10.3",
+		"@nestjs/common": "^7.3.2",
+		"@nestjs/core": "^7.3.2",
+		"@nestjs/platform-express": "^7.3.2",
+		"reflect-metadata": "^0.1.13"
+	},
+	"devDependencies": {
+		"@angular/cli": "10.0.1",
+		"@nrwl/workspace": "9.5.1",
+		"@types/node": "~14.0.20",
+		"dotenv": "8.2.0",
+		"ts-node": "~8.10.2",
+		"tslint": "~6.1.2",
+		"eslint": "7.4.0",
+		"typescript": "~3.9.6",
+		"prettier": "2.0.5",
+		"@nrwl/nest": "9.5.1",
+		"@angular/compiler-cli": "^10.0.3",
+		"@angular/language-service": "^10.0.3",
+		"@angular-devkit/build-angular": "0.1000.1",
+		"codelyzer": "~6.0.0",
+		"jest-preset-angular": "8.2.1",
+		"@nrwl/jest": "9.5.1",
+		"jest": "26.1.0",
+		"@types/jest": "26.0.4",
+		"ts-jest": "26.1.1",
+		"cypress": "^4.10.0",
+		"npm-check-updates": "~7.0.1",
+		"@nrwl/cypress": "9.5.1",
+		"@nrwl/node": "9.5.1",
+		"@nestjs/schematics": "^7.0.1",
+		"@nestjs/testing": "^7.3.2"
+	}
 }
