@@ -6,10 +6,10 @@ import { AppService } from './app.service';
 
 @Controller()
 export class AppController {
-  constructor(private readonly appService: AppService) {}
+	constructor(private readonly appService: AppService) {}
 
-  @Get('hello')
-  getData(): Message {
-    return this.appService.getData();
-  }
+	@Get('hello')
+	getData(): Message {
+		return this.appService.getData();
+	}
 }
