@@ -9,13 +9,13 @@ import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app/app.module';
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  const globalPrefix = 'api';
-  app.setGlobalPrefix(globalPrefix);
-  const port = process.env.PORT || 3333;
-  await app.listen(port, () => {
-    Logger.log('Listening at http://localhost:' + port + '/' + globalPrefix);
-  });
+	const app = await NestFactory.create(AppModule);
+	const globalPrefix = 'api';
+	app.setGlobalPrefix(globalPrefix);
+	const port = process.env.PORT || 3333;
+	await app.listen(port, () => {
+		Logger.log('Listening at http://localhost:' + port + '/' + globalPrefix);
+	});
 }
 
 bootstrap();
