@@ -1,13 +1,13 @@
 import { getGreeting } from '../support/app.po';
 
 describe('ngrx-demo', () => {
-  beforeEach(() => cy.visit('/'));
+	beforeEach(() => cy.visit('/'));
 
-  it('should display welcome message', () => {
-    // Custom command example, see `../support/commands.ts` file
-    cy.login('my-email@something.com', 'myPassword');
+	it('should display welcome message', () => {
+		// Custom command example, see `../support/commands.ts` file
+		cy.login('my-email@something.com', 'myPassword');
 
-    // Function helper example, see `../support/app.po.ts` file
-    getGreeting().contains('Welcome to ngrx-demo!');
-  });
+		// Function helper example, see `../support/app.po.ts` file
+		getGreeting().contains('Welcome to ngrx-demo!');
+	});
 });
