@@ -14,9 +14,9 @@
 const { preprocessTypescript } = require('@nrwl/cypress/plugins/preprocessor');
 
 module.exports = (on, config) => {
-  // `on` is used to hook into various events Cypress emits
-  // `config` is the resolved Cypress config
+	// `on` is used to hook into various events Cypress emits
+	// `config` is the resolved Cypress config
 
-  // Preprocess Typescript file using Nx helper
-  on('file:preprocessor', preprocessTypescript(config));
+	// Preprocess Typescript file using Nx helper
+	on('file:preprocessor', preprocessTypescript(config));
 };
