@@ -9,14 +9,14 @@
 // ***********************************************
 // eslint-disable-next-line @typescript-eslint/no-namespace
 declare namespace Cypress {
-  interface Chainable<Subject> {
-    login(email: string, password: string): void;
-  }
+	interface Chainable<Subject> {
+		login(email: string, password: string): void;
+	}
 }
 //
 // -- This is a parent command --
 Cypress.Commands.add('login', (email, password) => {
-  console.log('Custom command example: Login', email, password);
+	console.log('Custom command example: Login', email, password);
 });
 //
 // -- This is a child command --
