@@ -1,9 +1,9 @@
 module.exports = {
-  testMatch: ['**/+(*.)+(spec|test).+(ts|js)?(x)'],
-  transform: {
-    '^.+\\.(ts|js|html)$': 'ts-jest',
-  },
-  resolver: '@nrwl/jest/plugins/resolver',
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageReporters: ['html'],
+	testMatch: ['**/+(*.)+(spec|test).+(ts|js)?(x)'],
+	transform: {
+		'^.+\\.(ts|js|html)$': 'ts-jest',
+	},
+	resolver: '@nrwl/jest/plugins/resolver',
+	moduleFileExtensions: ['ts', 'js', 'html'],
+	coverageReporters: ['html'],
 };
