@@ -1,9 +1,9 @@
 module.exports = {
-  name: 'api-interfaces',
-  preset: '../../jest.config.js',
-  transform: {
-    '^.+\\.[tj]sx?$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-  coverageDirectory: '../../coverage/libs/api-interfaces',
+	name: 'api-interfaces',
+	preset: '../../jest.config.js',
+	transform: {
+		'^.+\\.[tj]sx?$': 'ts-jest',
+	},
+	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
+	coverageDirectory: '../../coverage/libs/api-interfaces',
 };
